serverless_function/network-controller/requirements.txt
python-kasa>=0.5.0
scapy>=2.5.0